# Perform a binary search to find optimal size of DNS responses