Will fix https://github.com/hashicorp/consul/issues/4036

Instead of removing one by one the entries, find the optimal
size using binary search.

For SRV records, with 5k nodes, duration of DNS lookups is
divided by 4 or more.

Results with around 5k nodes on local agent, dev mode:

Before Optimization: SRV~100ms ; A ~25ms
After Optimization: SRV ~20ms ; A ~20ms